Refactor: Place the request authorizer in the auth-adapter file
